backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
pydantic==2.10.4
packaging==24.2
